+++ Command/PurgeCommand.php
@@ -19,6 +19,9 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 
+/**
+ * Command to purge search indexes.
+ */
 class PurgeCommand extends Command
 {
     /**