+++ Tests/Unit/Search/ReIndex/Provider/DoctrineOrmProviderTest.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Massive\Bundle\SearchBundle\Tests\Unit\Search\ReIndex\Provider;
+namespace Massive\Bundle\SearchBundle\Tests\Unit\Search\Reindex\Provider;
 
 use Doctrine\Common\Persistence\Mapping\ClassMetadataFactory;
 use Doctrine\ORM\AbstractQuery;
@@ -19,7 +19,7 @@ use Doctrine\ORM\Mapping\ClassMetadata as OrmClassMetadata;
 use Doctrine\ORM\Query;
 use Doctrine\ORM\QueryBuilder;
 use Massive\Bundle\SearchBundle\Search\Metadata\ClassMetadata as SearchClassMetadata;
-use Massive\Bundle\SearchBundle\Search\ReIndex\Provider\DoctrineOrmProvider;
+use Massive\Bundle\SearchBundle\Search\Reindex\Provider\DoctrineOrmProvider;
 use Metadata\ClassHierarchyMetadata;
 use Metadata\MetadataFactory;