+++ Tests/Unit/Search/ReIndex/ResumeManagerTest.php
@@ -11,7 +11,7 @@
 
 namespace Massive\Bundle\SearchBundle\Tests\Unit\Search\Command;
 
-use Massive\Bundle\SearchBundle\Search\ReIndex\ResumeManager;
+use Massive\Bundle\SearchBundle\Search\Reindex\ResumeManager;
 use Symfony\Component\Filesystem\Filesystem;
 
 class ResumeManagerTest extends \PHPUnit_Framework_TestCase